feat: upgrade transporter to latest and fork from 4017700 (#190)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As an AVS developer, I want DevKit to automatically run a transporter
service that posts stake roots to the L1 chain at a frequency that I
specify, so that my aggregator and executor processes always have
accurate, up-to-date stake data for task validation.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Upgrades multichain-go to `v0.0.3`
- Moves fork block to `4017700`

**Result:**
<!--
*After your change, what will change.
-->
- Transporting roots works as expected

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Manually tested