feat: allow operators to define keys for each operatorSet (#246)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As a user of Devkit, I need to be able to define multiple keys for my
Operators so that I can assign the same operator to multiple
operatorSets.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Adds a `keystores` entry to the `operators` to list out all available
keys
- Adds a migration to move us to context version `v0.1.0`
- Removes remnant `skip-setup` context entry
- Adds tests for `v0.0.9` and `v0.1.0` context migrations

**Result:**
<!--
*After your change, what will change.
-->
- Users can assign multiple keys to their operators for different
operatorSets

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Migration tests have been added and other tests updated

**Open questions:**
<!--
(optional) Any open questions or feedback on design desired?
-->
- depends on:
https://github.com/Layr-Labs/hourglass-avs-template/pull/96